% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sternclean.R
\name{sternclean}
\alias{sternclean}
\title{A cleaner way to clean your data.}
\usage{
sternclean(the_data_frame = "the_data_frame", class_to_strng = "",
  class_to_numer = "", remove_columns = "", remove_na_rows = "",
  removeby_regex = "", remove_all_nas = FALSE, remove_non_num = FALSE,
  remove_all_exc = "", impute_na2mean = "", impute_na_cols = "",
  impute_na_with = "", impute_grpmean = "", impute_grpwith = "",
  impute_inf_col = "", impute_inf_wit = "", impute_cust_cl = "",
  impute_cust_fn = "", ...)
}
\arguments{
\item{the_data_frame}{tbl}

\item{class_to_strng}{vector of columns to convert to string.}

\item{class_to_numer}{vector of columns to convert to numeric.}

\item{remove_columns}{vector of columns to remove from tbl.}

\item{remove_na_rows}{vector of columns to remove rows by their NAs.}

\item{removeby_regex}{vector of columns found by regex to remove.}

\item{remove_all_nas}{TRUE/FALSE to remove all NAs by row. Default: False.}

\item{remove_non_num}{TRUE/FALSE to remove non-numeric columns. Default: False.}

\item{remove_all_exc}{vector of columns to keep. Remove all other columns.}

\item{impute_na2mean}{vector of columns to impute NAs with the column mean.}

\item{impute_na_cols}{vector of columns to impute with value given by the impute_na_with parameter.}

\item{impute_na_with}{value to impute in the vector of columnns from the impute_na_cols parameter.}

\item{impute_grpmean}{vector of columns to replace NAs with a groupmean, grouped by the impute_grpwith parameter.}

\item{impute_grpwith}{column to use as the group for the vector of columns defined by the impute_grpmean parameter.}

\item{impute_inf_col}{vector of columns to impute infinite values with the impute_inf_wit parameter value.}

\item{impute_inf_wit}{value to impute in the vector of columnns from the impute_inf_col parameter.}

\item{impute_cust_cl}{vector of columns to impute NAs with based on the function defined by the impute_cust_fn parameter.}

\item{impute_cust_fn}{function to impute the NAs over the columns defined in the impute_cust_cl parameter.}

\item{...}{additional parameters for the function from the impute_cust_fn parameter.}
}
\description{
A cleaner way to clean your data. It's recommended to keep all
parameters lined up and on separate lines to produce greater
readability and to make debugging easier. The function edits
the data frame and automatically assigns the results to the
data frames. We have kept the parameters starting with 3
separate verbs to make finding the correct parameter easier.
after typing either class, remove, or impute. One can simply
use tab to fill in the specific action they prefer.
}
\examples{
sternclean("airquality",
       impute_na2mean = c("Ozone", "Solar.R"))
sternclean("airquality",
       impute_na_cols = c("Ozone", "Solar.R"),
       impute_na_with = 1738)
sternclean("airquality",
       remove_columns = "Day"
       impute_na_cols = "Solar.R",
       impute_na_with = 1738)
}
